requirements.txt
streamlit
python-docx
pymupdf
python-dotenv
pydantic
bibtexparser
